requirements.txt
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-environ==0.4.54